move loading logic to respective JniLibrary classes
